LayoutTests/imported/w3c/web-platform-tests/2dcontext/transformations/2d.transformation.rotate.direction-expected.txt

2d.transformation.rotate.direction
rotate() is clockwise
Actual output:
Expected output:


PASS rotate() is clockwise 
